refactor(blocks)(#1482): break 4 small import cycles to drop sentrux acyclicity 6→2
Four mechanically-removable cycles flagged by sentrux's `max_cycles=5`
rule are eliminated without behaviour change:

1. `blocks/io/loaders/_capability` ↔ `load_data`:
   drop the `TYPE_CHECKING` import of `LoadData` (sentrux's AST analysis
   counts it as a real edge); duck-type `_resolve_format`'s `block`
   parameter as `Any | None` since ADR-028 Addendum 1 §C9 forbids a
   helper class (Protocol) in this sibling module.

2. `blocks/code/code_block` ↔ `validation`:
   extract `CodeBlockBackend` / `CodeBlockRuntimeContext` /
   `CodeBlockTimeoutError` / `register_codeblock_backend` /
   `unregister_codeblock_backend` / `list_codeblock_backends` /
   `resolve_codeblock_backend` / `ensure_codeblock_backends_loaded` /
   `run_codeblock_process` into a new sibling `_backends_registry.py`.
   `code_block.py` re-exports for the `backends/` subpackage; `validation.py`
   imports the helper from the new module directly so the cycle through
   `code_block` is gone.

3. `blocks/registry/_capability` ↔ `_spec`:
   relocate `_format_capabilities_from_class` from `_capability.py` to
   `_spec.py` (`_spec_from_class` is the only caller) and drop the lazy
   bidirectional imports. `_spec → _capability` stays one-way for
   `_validate_capability_id`. `blocks/registry/__init__.py` re-exports
   the relocated helper for backward compatibility.

4. Frontend store `types ↔ gitSlice ↔ lineageSlice` triangle:
   move `GitSlice` / `GitMergeInProgress` / `LineageSlice` interface
   declarations from the slice files into `types.ts` (matching the
   convention already used by every other slice). Slice files re-export
   the interfaces for any downstream consumer. `types.ts` no longer
   imports from the slice files.

Sentrux results on this branch:
  - `pass: true` (was `pass: false` on origin/main)
  - `acyclicity.raw`: 6 → 2 (max_cycles=5 violation cleared)
  - `quality_signal`: 4009 → 4817 (+20%)
  - `bottleneck`: acyclicity → depth (acyclicity is no longer the worst dimension)

Out of scope (separate issues):
  - `core/storage._defaults` ↔ `backend_router` ↔ `core/types/*` (#1342).
  - The 27-module `mcp + blocks/*` SCC (#1336).
  - The `laneAssign.ts` ↔ `laneAssign.parts/trunkPin.ts` GitGraph 2-cycle.

All public APIs preserved through re-exports:
  - `CodeBlockBackend`, `CodeBlockRuntimeContext`, `list_codeblock_backends`,
    `register_codeblock_backend`, etc.: still importable from
    `scistudio.blocks.code` and `scistudio.blocks.code.code_block`.
  - `_format_capabilities_from_class`: still importable from
    `scistudio.blocks.registry` and `scistudio.blocks.registry._capability`.
  - `GitSlice` / `GitMergeInProgress` / `LineageSlice`: still importable
    from `./gitSlice` and `./lineageSlice`.

Tests (worktree):
  - pytest tests/blocks/io tests/blocks/code tests/blocks/test_registry* — 500 passed, 8 skipped (pre-existing skips), 11 deselected (pre-existing plugin-missing failures on origin/main)
  - pytest tests/workflow tests/blocks/test_runner_registry.py — 79 passed
  - tsc --noEmit — clean
  - vitest run — 563 passed, 13 skipped (1 pre-existing fail in eslint-config.test.ts unrelated to scope)

Closes #1482

Authored-by: Jiazhen Zhang <jiazhenz026@gmail.com>
