+++ frontend/src/store/types.ts
@@ -2,6 +2,10 @@ import type {
   BlockSchemaResponse,
   BlockSummary,
   DataPreviewResponse,
+  GitBranch,
+  GitCommit,
+  GitHistoryFilter,
+  GitStatus,
   LogEntry,
   ProjectResponse,
   WorkflowEdge,
@@ -9,8 +13,16 @@ import type {
   WorkflowNode,
   WorkflowResponse,
 } from "../types/api";
+import type { LineageRunDetail, LineageRunSummary } from "../types/lineage";
 import type { BottomTab } from "../types/ui";
-import type { LineageSlice } from "./lineageSlice";
+
+// Issue #1482: ``GitSlice`` and ``LineageSlice`` interfaces previously
+// lived in their respective slice files and were re-imported here for
+// the ``AppStore`` union. Sentrux flagged the resulting
+// types ↔ gitSlice ↔ lineageSlice triangle as a cycle. The interfaces
+// now live here next to every other slice type
+// (ProjectSlice / WorkflowSlice / …); the slice files re-export them
+// for any downstream consumer.
 
 export interface ProjectDialogState {
   mode: "new" | "open";
@@ -416,10 +428,114 @@ export interface TabSlice {
   markFileRemoteConflict: (id: string, conflict: VersionConflictState) => void;
 }
 
-// ADR-039 §6 Phase 2 — git versioning slice
-// (defined in ./gitSlice.ts; imported here so AppStore unions it in)
-export type { GitSlice } from "./gitSlice";
-import type { GitSlice } from "./gitSlice";
+// ADR-039 §6 Phase 2 — git versioning slice (interfaces relocated here
+// from gitSlice.ts per issue #1482 to break the store ↔ types triangle).
+
+export interface GitMergeInProgress {
+  source_branch: string;
+  conflicted_files: string[];
+}
+
+export interface GitSlice {
+  branches: GitBranch[] | null;
+  currentBranch: string | null;
+  logCache: Record<string, GitCommit[]>;
+  logLoading: Record<string, boolean>;
+  historyFilter: GitHistoryFilter;
+  status: GitStatus | null;
+  mergeInProgress: GitMergeInProgress | null;
+  lastError: string | null;
+  /**
+   * ADR-039 Addendum 1 (#1354) — transient "safety auto-commit landed"
+   * notice. Set by `switchBranch` / `restore` when the backend
+   * response carries a non-null `auto_commit_sha`, consumed by
+   * `BranchPicker` (toast on switch) and `RestoreWorkflowButton`
+   * (inline hint on restore). Components clear it via
+   * `setLastNotice(null)` after rendering, mirroring the
+   * `lastError` lifecycle. Kept distinct from `lastError` so
+   * downstream UI does not confuse "your change was committed
+   * safely" with "something failed".
+   */
+  lastNotice: string | null;
+  /**
+   * ADR-039 §3.5 (#972 — Codex P1 on PR #974) — branch the user clicked
+   * "Merge into current" on. Driving this from the slice (rather than
+   * local Git-tab state) keeps the MergeFlow modal mounted at the
+   * BottomPanel level so switching bottom tabs during an in-flight
+   * conflict resolution does NOT tear it down and orphan the merge
+   * (MergeFlow's close guard would otherwise be bypassed). `null` =
+   * modal hidden.
+   */
+  mergeFlowSource: string | null;
+
+  /**
+   * Project ID active when `mergeFlowSource` was set (#975 Codex P1 on
+   * PR #980). Used by the App-level `<AppLevelMergeFlow>` mount to
+   * gate visibility: the modal renders only when the current open
+   * project matches this id. Switching to a different project hides
+   * the modal (state preserved); switching back re-shows it. Without
+   * this gate, modal actions like `complete merge` / `abort merge`
+   * would run against the wrong backend project context. `null` when
+   * no merge is in flight.
+   */
+  mergeFlowProjectId: string | null;
+
+  // Actions — D39-2.3b fills bodies.
+  setHistoryFilter: (filter: GitHistoryFilter) => void;
+  invalidateHistory: () => void;
+  loadBranches: () => Promise<void>;
+  loadLog: (branch?: string) => Promise<void>;
+  loadStatus: () => Promise<void>;
+  commit: (message: string, files?: string[]) => Promise<string>;
+  switchBranch: (name: string) => Promise<{ auto_commit_sha: string | null }>;
+  createBranch: (name: string, baseSha?: string) => Promise<void>;
+  deleteBranch: (name: string, force?: boolean) => Promise<void>;
+  restore: (
+    commitSha: string,
+    files?: string[],
+  ) => Promise<{ status: "ok"; auto_commit_sha: string | null }>;
+  setMergeInProgress: (state: GitMergeInProgress | null) => void;
+  /**
+   * Open or close MergeFlow. `source` is the branch being merged into
+   * the current branch (or `null` to close). `projectId` is the
+   * current open project's id — stamped here so the App-level mount
+   * can gate visibility against project switches (#975 Codex P1 on
+   * PR #980). Pass `null` for `projectId` when closing (`source=null`)
+   * or when opening outside any project context (test fixtures).
+   */
+  setMergeFlowSource: (source: string | null, projectId?: string | null) => void;
+  setLastError: (message: string | null) => void;
+  setLastNotice: (message: string | null) => void;
+}
+
+// ADR-038 §3.8 — Lineage tab client state (interface relocated here from
+// lineageSlice.ts per issue #1482).
+export interface LineageSlice {
+  // list pane
+  runs: LineageRunSummary[];
+  runsLoading: boolean;
+  runsError: string | null;
+  // detail pane
+  selectedRunId: string | null;
+  runDetails: Record<string, LineageRunDetail>;
+  runDetailLoading: Record<string, boolean>;
+  runDetailError: Record<string, string | null>;
+  // per-block expansion (UI-only)
+  expandedBlockExecutionIds: string[];
+  // dialogs (UI-only)
+  methodsDialogRunId: string | null;
+  rerunDialogRunId: string | null;
+  // actions
+  fetchRuns: (opts?: { workflowId?: string; limit?: number }) => Promise<void>;
+  fetchRunDetail: (runId: string) => Promise<void>;
+  selectRun: (runId: string | null) => void;
+  toggleBlockExecutionExpanded: (blockExecutionId: string) => void;
+  openMethodsDialog: (runId: string) => void;
+  closeMethodsDialog: () => void;
+  openRerunDialog: (runId: string) => void;
+  closeRerunDialog: () => void;
+  clearLineage: () => void;
+}
 
 export type AppStore = ProjectSlice &
   WorkflowSlice &
